requirements.txt
flask
Flask_RESTful
PyYAML
psycopg2